# Faster optional & nullable values