@FlorianWendelborn 2022-03-18
@tmcw shouldn’t this same optimization also be possible for [`.nullish()`](https://github.com/colinhacks/zod#nullish)?