@tmcw 2022-03-18
`nullish()` is a shortcut for `.optional().nullable()`, so this PR will increase performance somewhat for nullish as well - without changing how nullish works it isn't possible to optimize it much further.