@colinhacks 2022-03-18
Looks great yet again 🤙

The Deno thing is intentional - it gets built during a precommit hook into the `deno` subdir because I was too lazy to set up a separate repo and proper CI for it. I'll look into doing that in a `push` hook instead, I'm planning to overhaul the CI over the weekend.